fix dtype warning, test=develop

+++ lite/kernels/host/fill_constant_batch_size_like_compute.cc
@@ -46,7 +46,7 @@ void FillConstantBatchSizeLikeCompute::Run() {
     }
   } else if (param.dtype == static_cast<int32_t>(lite::core::FluidType::BOOL)) {
     auto data = param.out->template mutable_data<bool>();
-    for (int i = 0; i < param.out->numel(); i++) {
+    for (int64_t i = 0; i < param.out->numel(); i++) {
       data[i] = param.value;
     }
   } else {
